+++ modules/src/ics04_channel/handler/acknowledgement.rs
@@ -144,6 +144,7 @@ mod tests {
     use crate::mock::context::MockContext;
 
     use std::convert::TryFrom;
+    use std::time::Duration;
 
     #[test]
     fn ack_packet_processing() {
@@ -192,7 +193,7 @@ mod tests {
                 Default::default(),
             ),
             get_compatible_versions(),
-            0,
+            Duration::from_secs(0),
         );
 
         let tests: Vec<Test> = vec![