+++ relayer-cli/src/commands/tx/connection.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use abscissa_core::{Command, Options, Runnable};
 
 use ibc::events::IbcEvent;
@@ -56,7 +58,7 @@ impl Runnable for TxRawConnInitCmd {
             self,
             |chains: ChainHandlePair| {
                 Connection {
-                    delay_period: 0,
+                    delay_period: Duration::from_secs(0),
                     a_side: ConnectionSide::new(
                         chains.src,
                         self.src_client_id.clone(),
@@ -104,7 +106,7 @@ impl Runnable for TxRawConnTryCmd {
             self,
             |chains: ChainHandlePair| {
                 Connection {
-                    delay_period: 0,
+                    delay_period: Duration::from_secs(0),
                     a_side: ConnectionSide::new(
                         chains.src,
                         self.src_client_id.clone(),
@@ -160,7 +162,7 @@ impl Runnable for TxRawConnAckCmd {
             self,
             |chains: ChainHandlePair| {
                 Connection {
-                    delay_period: 0,
+                    delay_period: Duration::from_secs(0),
                     a_side: ConnectionSide::new(
                         chains.src,
                         self.src_client_id.clone(),
@@ -216,7 +218,7 @@ impl Runnable for TxRawConnConfirmCmd {
             self,
             |chains: ChainHandlePair| {
                 Connection {
-                    delay_period: 0,
+                    delay_period: Duration::from_secs(0),
                     a_side: ConnectionSide::new(
                         chains.src,
                         self.src_client_id.clone(),